project/app/requirements.txt
flask
flask-cors
pytesseract
pillow
opencv-python
